Added support for links.json
Due to irregular hyperlinks on the site, Iadded support for links.json to manually add pairs of currentLink:nextLink that the scraper will default to.

The main example of this is at the end of Volume 7, which I won't elaborate in detail due to spoilers.

I also cleaned up some comments as well